// Desktop/BharatGodam_App-divyang/src/types/entities.ts
export interface MongoDocument {
  _id: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}

interface price_perday {
  "_id": string,
  "price": number,
  "weight": string
}

// Define type for Commodity
export interface Commodity {
  name: string;
  bags: string;
  weight: string;
  AddedBy: string;
  price_perday: price_perday[];
  _id: string;
}
interface Address {
  street: string;
  landmark: string;
  pincode: string;
  city: string;
  state: string;
  buildingName: string;
}
// Define type for User
export interface User extends MongoDocument {
  Address: Address;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  role: string;
  isDocumentsVerified: boolean;
  isAddressVerified: boolean;
  isKycDone: boolean;
  lastLoginTime: string;
  isEmailVerified: boolean;
  isMobileVerified: boolean;
  isActive: boolean;
  isArchived: boolean;
  isVerified: boolean;
  isRejected: boolean;
  profilePicture: string;
  phone: string;
}
// ...
export interface WithdrawalDetails {
  User:string;
  commodity: {itemName: string; quantity: string}[];
  warehouse: {_id: string; warehouse_name: string};
  noOfBags: string;
  totalBags: string;
  truckNumber: string;
  driverName: string;
  bookingId: {
    _id: string;
    fromDate: string;
    toDate: string;
    noOfBags: string;
    bagSize: string;
  };
}

export interface addShipping {
  driverName: string;
  truckNumber: string;
  commodity: {itemName: string; quantity: string}[];
  totalBags: string;
}

export interface addShippingParams {
  warehouse_id: string;
  booking_id: string;
}